# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MuXiShooter",
        "title": "MuXiShooter",
        "termsOfService": "http://swagger.io/terms",
        "contact": {
            "name": "FireChickenMP4",
            "email": "13930176445@163.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/admin/get/resources": {
            "get": {
                "description": "通过query参数type查询skills/achievements/items/cards；支持分页与可选id精确查询",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-resource"
                ],
                "summary": "管理员按类型查询基础资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型(achievements/skills/items/cards)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "资源ID，传入后优先精确查询",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称模糊搜索",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "技能组模糊搜索(type=skills有效)",
                        "name": "skill_group",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页多少，默认20，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CommonAdminResourcePageData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "登录状态异常",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库查询失败",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/api/admin/get/user-relations": {
            "get": {
                "description": "通过query参数user_id和type查询指定用户在achievements/skills/items/cards中的关联数据\nskills会返回skill_grade，其他类型没有这个字段\ndata.list: []dto.CommonUserRelationData",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-resource"
                ],
                "summary": "管理员按类型查询任意用户关联数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "关联类型(achievements/skills/items/cards)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页多少，默认20，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CommonUserRelationPageData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "登录状态异常",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库查询失败",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/api/admin/operation/resources": {
            "post": {
                "description": "通过query参数type创建skills/achievements/items/cards中的一种资源\nskills需要额外参数skill_group和prq_skill_id，其他资源只需要公共请求体",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-resource"
                ],
                "summary": "管理员按类型创建基础资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型(achievements/skills/items/cards)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "创建请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CommonResourceCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CommonAdminResourceData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "登录状态异常",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "409": {
                        "description": "名称冲突",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "通过query参数type删除skills/achievements/items/cards中的一种资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-resource"
                ],
                "summary": "管理员按类型删除基础资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型(achievements/skills/items/cards)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "删除请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdminDeleteResourceByTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "登录状态异常",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "404": {
                        "description": "目标资源不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/api/admin/update/resources": {
            "put": {
                "description": "通过query参数type更新skills/achievements/items/cards中的一种资源\nskills需要额外参数skill_group和prq_skill_id，其他资源只需要公共请求体",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-resource"
                ],
                "summary": "管理员按类型更新基础资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型(achievements/skills/items/cards)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "更新请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CommonResourceUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CommonAdminResourceData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "登录状态异常",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "404": {
                        "description": "目标资源不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "409": {
                        "description": "名称冲突",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AdminDeleteResourceByTypeRequest": {
            "description": "用于skills/achievements/items/cards的删除（按ID）",
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CommonAdminResourceData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "prq_skill_id": {
                    "type": "integer"
                },
                "resource_id": {
                    "type": "integer"
                },
                "resource_name": {
                    "type": "string"
                },
                "skill_group": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CommonAdminResourcePageData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CommonAdminResourceData"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.CommonRelationResourceData": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "prq_skill_id": {
                    "type": "integer"
                },
                "resource_id": {
                    "type": "integer"
                },
                "resource_name": {
                    "type": "string"
                },
                "skill_group": {
                    "type": "string"
                }
            }
        },
        "dto.CommonResourceCreateRequest": {
            "description": "用于achievements/items/cards，skills可额外携带技能字段",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prq_skill_id": {
                    "type": "integer"
                },
                "skill_group": {
                    "type": "string"
                }
            }
        },
        "dto.CommonResourceUpdateRequest": {
            "description": "用于achievements/items/cards，skills可额外携带技能字段",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "prq_skill_id": {
                    "type": "integer"
                },
                "skill_group": {
                    "type": "string"
                }
            }
        },
...
        "dto.CommonUserRelationData": {
            "type": "object",
            "properties": {
                "claimed": {
                    "type": "boolean"
                },
                "claimed_at": {
                    "type": "string"
                },
                "complete_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "is_complete": {
                    "type": "boolean"
                },
                "resource": {
                    "$ref": "#/definitions/dto.CommonRelationResourceData"
                },
                "skill_grade": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CommonUserRelationPageData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CommonUserRelationData"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.Response": {
            "description": "通用响应结构体",
            "type": "object",
            "properties": {
                "code": {
                    "description": "http状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "根据具体数据来定类型"
                },
                "message": {
                    "description": "返回的消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入你的Bearer Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}